requirements.txt
scipy
opencv-python==3.4.2.16
matplotlib==2.2.4
Pillow==5.4.1
pytesseract==0.2.6
